+++ uninstall.sh
@@ -21,10 +21,11 @@ if [ -e ${User_servce} ]; then
     if [ "$enabled" = "enabled" ]; then
         sudo systemctl disable "$SERVICE_NAME"
     fi
+
+    if [ -e ${User_servce} ]; then
+        sudo rm -f /etc/systemd/system/dipper.service
+    fi
+    sudo systemctl daemon-reload
 fi
 
- # 以前のバージョン用
-sudo rm -f /etc/systemd/system/dipper.service
-
 sudo rm -rf /usr/local/dipper
-sudo systemctl daemon-reload
